refactor: create model
